chore: remove tsconfig for injected-scripts
