+++ .eslintignore
@@ -1,3 +1,4 @@
 .eslintrc.js
 *.config.js
 tsconfig*.json
+*/.eslintrc.js
