+++ core/injected-scripts/.eslintrc.js
@@ -1,10 +1,5 @@
-const { join } = require('path');
-
 module.exports = {
   extends: '../../.eslintrc.js',
-  parserOptions: {
-    project: join(__dirname, 'tsconfig.json'),
-  },
   overrides: [
     {
       files: ['**/*.ts'],
