+++ core/injected-scripts/pageEventsRecorder.ts
@@ -6,7 +6,15 @@ import type { IFocusEvent } from '@ulixee/hero-interfaces/IFocusEvent';
 import type { IScrollEvent } from '@ulixee/hero-interfaces/IScrollEvent';
 import type { ILoadEvent } from '@ulixee/hero-interfaces/ILoadEvent';
 
+interface ITypeSerializer {
+  stringify(object: any): string;
+  parse(object: string): any;
+  replace(object: any): any;
+}
+
 declare global {
+  let TypeSerializer: ITypeSerializer;
+
   interface Window {
     extractDomChanges(): PageRecorderResultSet;
     flushPageRecorder(): PageRecorderResultSet;
