--- core/injected-scripts/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "composite": true,
-    "sourceMap": false,
-    "inlineSourceMap": false,
-    "removeComments": true
-  },
-  "include": ["*.ts", ".eslintrc.js"],
-  "exclude": ["**/tsconfig*.json", "**/node_modules", "**/build"]
-}
