--- timetravel/injected-scripts/tsconfig.dist.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../../tsconfig.dist.json",
-  "compilerOptions": {
-    "composite": true,
-    "sourceMap": false,
-    "inlineSourceMap": false,
-    "removeComments": true
-  },
-  "include": ["*.ts", ".eslintrc.js"],
-  "exclude": ["**/tsconfig*.json", "**/node_modules", "**/build-dist"]
-}
