+++ tsconfig.json
@@ -5,14 +5,6 @@
     "outDir": "./build",
     "noImplicitOverride": true
   },
-  "references": [
-    {
-      "path": "core/injected-scripts/tsconfig.json"
-    },
-    {
-      "path": "timetravel/injected-scripts/tsconfig.json"
-    }
-  ],
   "include": ["./**/*.ts", "**/*.js", "**/*.json", ".eslintrc.js"],
   "exclude": [
     "**/tsconfig*.json",
@@ -23,8 +15,6 @@
     "build",
     "build-dist",
     "unblocked",
-    "core/injected-scripts",
-    "timetravel/injected-scripts",
     "examples/*.js",
     "**/*.d.ts"
   ]
